+++ core/modules/block/block.routing.yml
@@ -84,3 +84,12 @@ block.category_autocomplete:
     _controller: '\Drupal\block\Controller\CategoryAutocompleteController::autocomplete'
   requirements:
     _permission: 'administer blocks'
+
+entity.block.off_canvas_form:
+  path: '/admin/structure/block/manage/{block}/off-canvas'
+  defaults:
+    _entity_form: 'block.off_canvas'
+    _title_callback: '\Drupal\block\Form\BlockEntityOffCanvasForm::title'
+  requirements:
+    _permission: 'administer blocks'
+
