--- core/modules/outside_in/outside_in.links.contextual.yml
@@ -1,4 +1,0 @@
-outside_in.block_configure:
-  title: 'Quick edit'
-  route_name: 'entity.block.off_canvas_form'
-  group: 'block'
